EST: extract ServletContextListener
The recent commit f4aeb60346d59d10d2ae3b213ba407b4df9dbeaa extracted
the ServletContextListener implementation from the "engine" class
for each existing subsystem.  This commit does likewise for the EST
subsystem.

Part of: https://github.com/dogtagpki/pki/issues/3297
